move rosrun and roslaunch functor to separate file
