+++ src/extension.ts
@@ -14,6 +14,9 @@ import * as pfs from "./promise-fs";
 import * as utils from "./utils";
 import * as telemetry from "./telemetry";
 
+import * as rosrun from "./ros/rosrun";
+import * as roslaunch from "./ros/roslaunch";
+
 /**
  * The catkin workspace base dir.
  */
@@ -179,72 +182,15 @@ function activateEnvironment(context: vscode.ExtensionContext) {
             build.updatePythonPath(logger);
         }),
         vscode.commands.registerCommand(Commands.Rosrun, () => {
-            rosrundelegate(logger);
+            rosrun.setup(logger);
         }),
         vscode.commands.registerCommand(Commands.Roslaunch, () => {
-            roslaunchdelegate(logger);
+            roslaunch.setup(logger);
         }),
     );
 
     // Generate config files if they don't already exist.
     build.createConfigFiles();
-}
-
-async function rosrundelegate(logger?: telemetry.ILogger) {
-    if (logger) {
-        logger.logCommand(Commands.Rosrun);
-    }
-
-    let terminal = await preparerosrun();
-    terminal.show();
-}
-
-async function preparerosrun(): Promise<vscode.Terminal> {
-    const packages = utils.getPackages();
-    const packageName = await vscode.window.showQuickPick(packages.then(Object.keys), { placeHolder: "Choose a package" });
-    if (packageName !== undefined) {
-        let basenames = (files: string[]) => files.map(file => path.basename(file));
-
-        const executables = utils.findPackageExecutables(packageName).then(basenames);
-        let target = await vscode.window.showQuickPick(executables, { placeHolder: "Choose an executable" });
-        let argument = await vscode.window.showInputBox({ placeHolder: "Enter any extra arguments" });
-        let terminal = vscode.window.createTerminal({
-            name: "rosrun",
-            env: env
-        });
-        terminal.sendText(`rosrun ${packageName} ${target} ${argument}`);
-        return terminal;
-    } else {
-        // none of the packages selected, error!
-    }
-}
-
-async function roslaunchdelegate(logger?: telemetry.ILogger) {
-    if (logger) {
-        logger.logCommand(Commands.Roslaunch);
-    }
-
-    let terminal = await prepareroslaunch();
-    terminal.show();
-}
-
-async function prepareroslaunch(): Promise<vscode.Terminal> {
-    const packages = utils.getPackages();
-    const packageName = await vscode.window.showQuickPick(packages.then(Object.keys), { placeHolder: "Choose a package" });
-    if (packageName !== undefined) {
-        const launchFiles = await utils.findPackageLaunchFiles(packageName);
-        const launchFileBasenames = launchFiles.map((filename) => path.basename(filename));
-        let target = await vscode.window.showQuickPick(launchFileBasenames, { placeHolder: "Choose a launch file" });
-        let argument = await vscode.window.showInputBox({ placeHolder: "Enter any extra arguments" });
-        let terminal = vscode.window.createTerminal({
-            name: "roslaunch",
-            env: env
-        });
-        terminal.sendText(`roslaunch ${launchFiles[launchFileBasenames.indexOf(target)]} ${argument}`);
-        return terminal;
-    } else {
-        // none of the packages selected, error!
-    }
 }
 
 /**
